requirements.txt
Flask==3.0.0
flask-cors==4.0.0
g4f==0.2.8.8
gunicorn==21.2.0
aiohttp==3.9.1
requests==2.31.0
discord.py==2.3.2
python-dotenv==1.0.0
